Split func. between .ex modules, cached displayable post
