+++ lib/blog_web/templates/page/posts.html.heex
@@ -6,6 +6,6 @@
 <.post_previews page={@page}/>
 
 
-<%= if Blog.Posts.page_exists?(@page + 1) do %>
+<%= if Blog.Posts.valid_post_page_exists?(@page + 1, posts_per_page: 1) do %>
   <div><.link href={"/posts/?page=" <> Integer.to_string(@page + 1) }>Next page</.link></div>
 <% end %>
