+++ lib/blog_web/views/page_view.ex
@@ -116,27 +116,9 @@ defmodule BlogWeb.PageView do
   end
 
   def post_previews(assigns) do
-    valid_posts = Blog.Posts.valid_posts()
-
     posts_per_page = 1
-
     page = assigns[:page] || 1
-
-    posts =
-      (valid_posts
-       |> Enum.sort_by(fn post -> post.dt end, {:desc, NaiveDateTime})
-       |> Stream.chunk_every(posts_per_page)
-       |> Enum.at(page - 1) || [])
-      |> Enum.map(fn post ->
-        Map.take(post.manifest.parsed, [:title, :description, :tags])
-        |> Map.put(
-          :date,
-          case Timex.format(post.dt, "{Mshort} {D}, {YYYY}") do
-            {:ok, fmt} -> fmt
-          end
-        )
-        |> Map.put(:id, post.id)
-      end)
+    posts = Blog.Posts.posts_for_display() |> Blog.Posts.take_page(page, posts_per_page)
 
     assigns = Map.put(assigns, :posts, posts)
